requirements.txt
netifaces
iptools
pings
dnspython
pyfiglet